Redesigned how photo modal component operates/displays
No longer routes the user to the the modal component when cliking a photo. This was unmounting and remounting the user profile component every time you clicked on a photo in the past. Now it will display modal over the current page and change the history path/url. If user reloads the page, they will be brought to the specific photo show page which corresponds to the route.